[DEBT-043-D12] Split portfolio/syncer.py into focused modules
- Split 483 LoC into 6 files under 400 lines each:
  - _sync_helpers.py (81): compute_fifo_cost_basis, SyncResult
  - _sync_positions.py (130): Position sync logic
  - _sync_trades.py (116): Trade sync logic
  - _sync_settlements.py (99): Settlement sync logic
  - _mark_prices.py (112): Mark price update logic
  - syncer.py (100): PortfolioSyncer orchestrator class
- Preserved backwards-compatible imports via syncer.py re-exports
- All tests pass: 1052 passed
- Quality gates: pre-commit, mypy, pytest
