+++ PROGRESS.md
@@ -26,7 +26,7 @@
 - [x] **DEBT-043-D9**: Split `exa/client.py` into focused modules → `docs/_debt/DEBT-043-srp-god-files.md`
 - [x] **DEBT-043-D10**: Split `exa/websets/client.py` into focused modules → `docs/_debt/DEBT-043-srp-god-files.md`
 - [x] **DEBT-043-D11**: Split `portfolio/pnl.py` into focused modules → `docs/_debt/DEBT-043-srp-god-files.md`
-- [ ] **DEBT-043-D12**: Split `portfolio/syncer.py` into focused modules → `docs/_debt/DEBT-043-srp-god-files.md`
+- [x] **DEBT-043-D12**: Split `portfolio/syncer.py` into focused modules → `docs/_debt/DEBT-043-srp-god-files.md`
 - [ ] **DEBT-043-D13**: Split `analysis/liquidity.py` into focused modules → `docs/_debt/DEBT-043-srp-god-files.md`
 - [ ] **DEBT-043-D14**: Split `analysis/scanner.py` into focused modules → `docs/_debt/DEBT-043-srp-god-files.md`
 - [ ] **DEBT-043-D15**: Split `analysis/correlation.py` into focused modules → `docs/_debt/DEBT-043-srp-god-files.md`
@@ -42,6 +42,7 @@
 
 ## Work Log
 
+- 2026-01-21: **DEBT-043-D12** complete — split `portfolio/syncer.py` (483 LoC) into 6 files: `_sync_helpers.py` (81), `_sync_positions.py` (130), `_sync_trades.py` (116), `_sync_settlements.py` (99), `_mark_prices.py` (112), `syncer.py` (100). `PortfolioSyncer` delegates to module functions while preserving backwards-compatible class interface. Quality gates: `pre-commit`, `pytest` (1052 passed).
 - 2026-01-21: **DEBT-043-D11** complete — split `portfolio/pnl.py` (554 LoC) into 4 files: `_pnl_models.py` (80), `_fifo.py` (174), `_settlements.py` (191), `pnl.py` (276). `PnLCalculator` delegates to module functions while preserving backwards-compatible class interface. Quality gates: `pre-commit`, `pytest` (1052 passed).
 - 2026-01-21: **DEBT-043-D10** complete — split `exa/websets/client.py` (442 LoC) into mixins: `_http.py` (219), `_websets.py` (117), `_items.py` (80), `_searches.py` (100), `client.py` (54). Used mixin composition via multiple inheritance. Quality gates: `pre-commit`, `pytest` (1052 passed).
 - 2026-01-21: **DEBT-043-D9** complete — split `exa/client.py` (694 LoC) into mixins: `_http.py` (228), `_normalization.py` (50), `_search.py` (262), `_contents.py` (100), `_answer.py` (52), `_research.py` (178), `client.py` (54). Updated test patches to use new module paths. Quality gates: `pre-commit`, `pytest` (1052 passed).
